requirements.txt
# Synexis Rep Agent — Python Dependencies

# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Vector store
pinecone==5.0.1

# Embeddings + reranking
voyageai==0.3.2

# LLM
anthropic==0.34.2

# Content pipeline — extraction
pdfplumber==0.11.4        # PDF text + table extraction
pymupdf==1.24.10          # Fallback PDF extraction (handles more encoding edge cases)
python-docx==1.1.2        # docx extraction
python-pptx==1.0.2        # pptx extraction

# Content pipeline — chunking + tokenization
tiktoken==0.7.0           # Token counting for chunk size targeting

# Retrieval — sparse (BM25)
rank-bm25==0.2.2

# Utilities
python-dotenv==1.0.1
httpx==0.27.2             # Async HTTP client
tenacity==8.5.0           # Retry logic for API calls
tqdm==4.66.5              # Progress bars for pipeline runs
